occlusion_test

+++ vit_att_trial/occlusion_test.py
@@ -90,7 +90,7 @@ def occlusion(model, image, label, occ_size=100, occ_stride=100, occ_pixel=0.5):
     # vis = np.uint8(255 * vis)
     # vis = cv2.cvtColor(np.array(vis), cv2.COLOR_RGB2BGR)
     heatmap = heatmap.permute(1, 0)
-    new_heatmap =cv2.resize(heatmap, (512,496))
+    new_heatmap =cv2.resize(np.float32(heatmap), (512,496))
 
     # heatmap = (heatmap - heatmap.min()) / (
     #         heatmap.max() - heatmap.min())
